[#148484799] 🚜 (Components) Update file structure for remaining components
