+++ app/components/List/Ul.test.js
@@ -1,29 +1,30 @@
 import React from 'react'
 import { render } from '@testing-library/react'
 
-import Wrapper from '../Wrapper'
+import Ul from './Ul'
 
-describe('<Wrapper />', () => {
-  it('should render an <div> tag', () => {
-    const { container } = render(<Wrapper />)
-    expect(container.firstElementChild.tagName).toEqual('DIV')
+describe('<Ul />', () => {
+  it('should render an <ul> tag', () => {
+    const { container } = render(<Ul />)
+    const element = container.firstElementChild
+    expect(element.tagName).toEqual('UL')
   })
 
   it('should have a class attribute', () => {
-    const { container } = render(<Wrapper />)
+    const { container } = render(<Ul />)
     const element = container.firstElementChild
     expect(element.hasAttribute('class')).toBe(true)
   })
 
   it('should adopt a valid attribute', () => {
     const id = 'test'
-    const { container } = render(<Wrapper id={id} />)
+    const { container } = render(<Ul id={id} />)
     const element = container.firstElementChild
     expect(element.id).toEqual(id)
   })
 
   it('should not adopt an invalid attribute', () => {
-    const { container } = render(<Wrapper attribute="test" />)
+    const { container } = render(<Ul attribute="test" />)
     const element = container.firstElementChild
     expect(element.hasAttribute('attribute')).toBe(false)
   })
