+++ app/components/ListItem/ListItem.js
@@ -1,0 +1,19 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+import Item from './Item'
+import Wrapper from './Wrapper'
+
+function ListItem(props) {
+  return (
+    <Wrapper>
+      <Item>{props.item}</Item>
+    </Wrapper>
+  )
+}
+
+ListItem.propTypes = {
+  item: PropTypes.any,
+}
+
+export default ListItem
