+++ app/components/ListItem/Wrapper.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render } from '@testing-library/react'
 
-import Wrapper from '../Wrapper'
+import Wrapper from './Wrapper'
 
 describe('<Wrapper />', () => {
   it('should render an <li> tag', () => {
