+++ app/components/ListItem/index.js
@@ -1,19 +1,1 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-
-import Item from './Item'
-import Wrapper from './Wrapper'
-
-function ListItem(props) {
-  return (
-    <Wrapper>
-      <Item>{props.item}</Item>
-    </Wrapper>
-  )
-}
-
-ListItem.propTypes = {
-  item: PropTypes.any,
-}
-
-export default ListItem
+export { default as ListItem } from './ListItem'
