+++ app/components/LoadingIndicator/LoadingIndicator.test.js
@@ -2,7 +2,7 @@ import React from 'react'
 import renderer from 'react-test-renderer'
 import 'jest-styled-components'
 
-import LoadingIndicator from '../index'
+import LoadingIndicator from './LoadingIndicator'
 
 describe('<LoadingIndicator />', () => {
   it('should match the snapshot', () => {
