+++ app/components/Toggle/Select.test.js
@@ -1,28 +1,27 @@
 import React from 'react'
 import { render } from '@testing-library/react'
 
-import Item from '../Item'
+import Select from './Select'
 
-describe('<Item />', () => {
-  it('should render an <div> tag', () => {
-    const { container } = render(<Item />)
-    expect(container.firstChild.tagName).toEqual('DIV')
+describe('<Select />', () => {
+  it('should render an <select> tag', () => {
+    const { container } = render(<Select />)
+    expect(container.firstChild.tagName).toEqual('SELECT')
   })
 
   it('should have a class attribute', () => {
-    const { container } = render(<Item />)
+    const { container } = render(<Select />)
     expect(container.firstChild.hasAttribute('class')).toBe(true)
   })
 
   it('should adopt a valid attribute', () => {
     const id = 'test'
-    const { container } = render(<Item id={id} />)
-    expect(container.firstChild.hasAttribute('id')).toBe(true)
+    const { container } = render(<Select id={id} />)
     expect(container.firstChild.id).toEqual(id)
   })
 
   it('should not adopt an invalid attribute', () => {
-    const { container } = render(<Item attribute="test" />)
+    const { container } = render(<Select attribute="test" />)
     expect(container.firstChild.hasAttribute('attribute')).toBe(false)
   })
 })
