+++ app/containers/App/App.js
@@ -15,7 +15,7 @@ import HomePage from 'containers/HomePage/HomePage.loadable'
 import FeaturePage from 'containers/FeaturePage/FeaturePage.loadable'
 import NotFoundPage from 'containers/NotFoundPage/NotFoundPage.loadable'
 import { TemplateHeader } from 'components/TemplateHeader'
-import Footer from 'components/Footer'
+import { Footer } from 'components/Footer'
 
 import GlobalStyle from '../../global-styles'
