+++ app/containers/FeaturePage/FeaturePage.loadable.js
@@ -4,7 +4,7 @@
 
 import React from 'react'
 import loadable from 'utils/loadable'
-import LoadingIndicator from 'components/LoadingIndicator'
+import { LoadingIndicator } from 'components/LoadingIndicator'
 
 export default loadable(() => import('./FeaturePage'), {
   fallback: <LoadingIndicator />,
